refactor: extract secrets core package
