+++ packages/plugin-sdk/tsconfig.json
@@ -16,6 +16,7 @@
     "../../packages/markdown-core/src/**/*.ts",
     "../../packages/media-generation-core/src/**/*.ts",
     "../../packages/model-catalog-core/src/**/*.ts",
+    "../../packages/secrets-core/src/**/*.ts",
     "../../packages/terminal-core/src/**/*.ts",
     "../../src/plugin-sdk/**/*.ts",
     "../../src/video-generation/dashscope-compatible.ts",
