+++ packages/secrets-core/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./secret-input.js";
+export * from "./secret-ref.js";
