+++ packages/secrets-core/src/secret-ref.ts
@@ -1,0 +1,163 @@
+export type SecretRefSource = "env" | "file" | "exec"; // pragma: allowlist secret
+
+/**
+ * Stable identifier for a secret in a configured source.
+ * Examples:
+ * - env source: provider "default", id "OPENAI_API_KEY"
+ * - file source: provider "mounted-json", id "/providers/openai/apiKey"
+ * - exec source: provider "vault", id "openai/api-key"
+ */
+export type SecretRef = {
+  source: SecretRefSource;
+  provider: string;
+  id: string;
+};
+
+export type SecretInput = string | SecretRef;
+export type SecretDefaults = {
+  env?: string;
+  file?: string;
+  exec?: string;
+};
+
+export const DEFAULT_SECRET_PROVIDER_ALIAS = "default"; // pragma: allowlist secret
+export const ENV_SECRET_REF_ID_RE = /^[A-Z][A-Z0-9_]{0,127}$/;
+export const LEGACY_SECRETREF_ENV_MARKER_PREFIX = "secretref-env:"; // pragma: allowlist secret
+export const LEGACY_DOUBLE_UNDERSCORE_ENV_MARKER_PREFIX = "__env__:"; // pragma: allowlist secret
+
+const ENV_SECRET_TEMPLATE_RE = /^\$\{([A-Z][A-Z0-9_]{0,127})\}$/;
+const ENV_SECRET_SHORTHAND_RE = /^\$([A-Z][A-Z0-9_]{0,127})$/;
+const SECRET_REF_SOURCES = new Set<SecretRefSource>(["env", "file", "exec"]);
+
+function isRecord(value: unknown): value is Record<string, unknown> {
+  return typeof value === "object" && value !== null && !Array.isArray(value);
+}
+
+function hasSecretRefSource(value: unknown): value is SecretRefSource {
+  return typeof value === "string" && SECRET_REF_SOURCES.has(value as SecretRefSource);
+}
+
+function hasNonEmptyString(value: unknown): value is string {
+  return typeof value === "string" && value.trim().length > 0;
+}
+
+export function isValidEnvSecretRefId(value: string): boolean {
+  return ENV_SECRET_REF_ID_RE.test(value);
+}
+
+export function isSecretRef(value: unknown): value is SecretRef {
+  if (!isRecord(value)) {
+    return false;
+  }
+  if (Object.keys(value).length !== 3) {
+    return false;
+  }
+  return (
+    hasSecretRefSource(value.source) &&
+    hasNonEmptyString(value.provider) &&
+    hasNonEmptyString(value.id)
+  );
+}
+
+function isLegacySecretRefWithoutProvider(
+  value: unknown,
+): value is { source: SecretRefSource; id: string } {
+  if (!isRecord(value)) {
+    return false;
+  }
+  return (
+    hasSecretRefSource(value.source) && hasNonEmptyString(value.id) && value.provider === undefined
+  );
+}
+
+export function parseEnvTemplateSecretRef(
+  value: unknown,
+  provider = DEFAULT_SECRET_PROVIDER_ALIAS,
+): SecretRef | null {
+  if (typeof value !== "string") {
+    return null;
+  }
+  const trimmed = value.trim();
+  const match = ENV_SECRET_TEMPLATE_RE.exec(trimmed) ?? ENV_SECRET_SHORTHAND_RE.exec(trimmed);
+  if (!match) {
+    return null;
+  }
+  return {
+    source: "env",
+    provider: provider.trim() || DEFAULT_SECRET_PROVIDER_ALIAS,
+    id: match[1],
+  };
+}
+
+export function parseLegacySecretRefEnvMarker(
+  value: unknown,
+  provider = DEFAULT_SECRET_PROVIDER_ALIAS,
+): SecretRef | null {
+  if (typeof value !== "string") {
+    return null;
+  }
+  const trimmed = value.trim();
+  const prefix = trimmed.startsWith(LEGACY_SECRETREF_ENV_MARKER_PREFIX)
+    ? LEGACY_SECRETREF_ENV_MARKER_PREFIX
+    : trimmed.startsWith(LEGACY_DOUBLE_UNDERSCORE_ENV_MARKER_PREFIX)
+      ? LEGACY_DOUBLE_UNDERSCORE_ENV_MARKER_PREFIX
+      : undefined;
+  if (!prefix) {
+    return null;
+  }
+  const id = trimmed.slice(prefix.length);
+  if (!ENV_SECRET_REF_ID_RE.test(id)) {
+    return null;
+  }
+  return {
+    source: "env",
+    provider: provider.trim() || DEFAULT_SECRET_PROVIDER_ALIAS,
+    id,
+  };
+}
+
+export function coerceSecretRef(value: unknown, defaults?: SecretDefaults): SecretRef | null {
+  if (isSecretRef(value)) {
+    return value;
+  }
+  const legacyEnvMarker = parseLegacySecretRefEnvMarker(value, defaults?.env);
+  if (legacyEnvMarker) {
+    return legacyEnvMarker;
+  }
+  if (isLegacySecretRefWithoutProvider(value)) {
+    const provider =
+      value.source === "env"
+        ? (defaults?.env ?? DEFAULT_SECRET_PROVIDER_ALIAS)
+        : value.source === "file"
+          ? (defaults?.file ?? DEFAULT_SECRET_PROVIDER_ALIAS)
+          : (defaults?.exec ?? DEFAULT_SECRET_PROVIDER_ALIAS);
+    return {
+      source: value.source,
+      provider,
+      id: value.id,
+    };
+  }
+  const envTemplate = parseEnvTemplateSecretRef(value, defaults?.env);
+  if (envTemplate) {
+    return envTemplate;
+  }
+  return null;
+}
+
+export function resolveSecretInputRef(params: {
+  value: unknown;
+  refValue?: unknown;
+  defaults?: SecretDefaults;
+}): {
+  explicitRef: SecretRef | null;
+  inlineRef: SecretRef | null;
+  ref: SecretRef | null;
+} {
+  const explicitRef = coerceSecretRef(params.refValue, params.defaults);
+  const inlineRef = explicitRef ? null : coerceSecretRef(params.value, params.defaults);
+  return {
+    explicitRef,
+    inlineRef,
+    ref: explicitRef ?? inlineRef,
+  };
+}
