+++ scripts/build-all.mjs
@@ -51,6 +51,7 @@ export const BUILD_ALL_STEPS = [
         "packages/media-understanding-common/package.json",
         "packages/terminal-core/package.json",
         "packages/model-catalog-core/package.json",
+        "packages/secrets-core/package.json",
         "packages/web-content-core/package.json",
         "packages/memory-host-sdk/package.json",
         "tsconfig.json",
@@ -59,6 +60,7 @@ export const BUILD_ALL_STEPS = [
         "packages/llm-core/src",
         "packages/markdown-core/src",
         "packages/model-catalog-core/src",
+        "packages/secrets-core/src",
         "packages/memory-host-sdk/src",
         "packages/media-generation-core/src",
         "packages/media-understanding-common/src",
