+++ scripts/prepare-extension-package-boundary-artifacts.mjs
@@ -18,6 +18,7 @@ const PLUGIN_SDK_TYPE_INPUTS = [
   "packages/llm-core/src",
   "packages/markdown-core/src",
   "packages/model-catalog-core/src",
+  "packages/secrets-core/src",
   "packages/memory-host-sdk/src",
   "packages/media-generation-core/src",
   "packages/media-understanding-common/src",
@@ -76,6 +77,9 @@ const ROOT_DTS_REQUIRED_OUTPUTS = [
   "dist/plugin-sdk/packages/model-catalog-core/src/provider-id.d.ts",
   "dist/plugin-sdk/packages/model-catalog-core/src/provider-model-id-normalization.d.ts",
   "dist/plugin-sdk/packages/model-catalog-core/src/provider-model-id-normalize.d.ts",
+  "dist/plugin-sdk/packages/secrets-core/src/index.d.ts",
+  "dist/plugin-sdk/packages/secrets-core/src/secret-input.d.ts",
+  "dist/plugin-sdk/packages/secrets-core/src/secret-ref.d.ts",
   "dist/plugin-sdk/error-runtime.d.ts",
   "dist/plugin-sdk/plugin-entry.d.ts",
   "dist/plugin-sdk/provider-auth.d.ts",
@@ -105,6 +109,9 @@ const PACKAGE_DTS_REQUIRED_OUTPUTS = [
   "packages/plugin-sdk/dist/packages/model-catalog-core/src/provider-id.d.ts",
   "packages/plugin-sdk/dist/packages/model-catalog-core/src/provider-model-id-normalization.d.ts",
   "packages/plugin-sdk/dist/packages/model-catalog-core/src/provider-model-id-normalize.d.ts",
+  "packages/plugin-sdk/dist/packages/secrets-core/src/index.d.ts",
+  "packages/plugin-sdk/dist/packages/secrets-core/src/secret-input.d.ts",
+  "packages/plugin-sdk/dist/packages/secrets-core/src/secret-ref.d.ts",
   "packages/plugin-sdk/dist/packages/terminal-core/src/ansi.d.ts",
   "packages/plugin-sdk/dist/packages/terminal-core/src/decorative-emoji.d.ts",
   "packages/plugin-sdk/dist/packages/terminal-core/src/health-style.d.ts",
