+++ src/plugins/contracts/extension-package-project-boundaries.test.ts
@@ -197,6 +197,7 @@ describe("opt-in extension package boundaries", () => {
       "../../packages/markdown-core/src/**/*.ts",
       "../../packages/media-generation-core/src/**/*.ts",
       "../../packages/model-catalog-core/src/**/*.ts",
+      "../../packages/secrets-core/src/**/*.ts",
       "../../packages/terminal-core/src/**/*.ts",
       "../../src/plugin-sdk/**/*.ts",
       "../../src/video-generation/dashscope-compatible.ts",
