+++ tsconfig.plugin-sdk.dts.json
@@ -18,6 +18,7 @@
     "packages/media-generation-core/src/**/*.ts",
     "packages/model-catalog-core/src/**/*.ts",
     "packages/memory-host-sdk/src/**/*.ts",
+    "packages/secrets-core/src/**/*.ts",
     "packages/terminal-core/src/**/*.ts",
     "src/video-generation/dashscope-compatible.ts",
     "src/video-generation/types.ts",
